Updated SDBManager and query to take an optional "quick" parameter which now specifies if we should do a full count, or just return the first result set. This is due to Amazon's definition of count(*):
If the count request takes more than five seconds, Amazon SimpleDB
returns the number of items that it could count and a next token to
return additional results. The client is responsible for accumulating
the partial counts.

But in the general case, we don't want to take a long time returning the
count, we just want a quick result. Quick is set to "True" by default,
but you can get a FULL count by doing:

    query.count(False)