Fixed the issue of "None" values in SDB. If you set a value to None" in an item, it would set that value to the string "None" instead of removing that attribute.
Now, it will actually make a second call to delete any attributes
that are set to None.

Also, the SDB manager for some reason doesn't use the item directly,
but instead uses the domain. I added the same functionality there,
and removed all the weird work-arounds that set things to "None"
instead of removing them. This makes the "is null" comparison work
again.