Rolled back the APIVersion for the ec2 connection so we get instance tags.
For reasons yet unknown, the more current version specification causes the XML persing for get_all_instances to ignore anything between a close item tag (</item>) and the next open item tag. This causes the code not to process the tag information, so we don't get instance tags.

This fix is a necessary temporary step for our use of boto. A permanent fix is obviously required so the interface version can be upgraded.