Fix first SQS test.
The first SQS test is to try to create a queue with an illegal name,
but the name used isn't actually illegal.  The first time you run the
test, it create a queue with that name, which then throws the queue
count off for the next test.

The name used now includes a "*" which is illegal, and also will fail
if no exception is thrown when creating the queue.