Updated the Table class to return 0 when ItemCount or TableSizeBytes is missing
The response from AWS upon creating a new table does not include the
'Table' dictionary, so the ItemCount and TableSizeBytes values do not
get added to the _dict attribute immediately after the creation of a
new table. This causes calls to the item_count and size_bytes
properties to return KeyErrors.

This fix returns a zero for both attributes when either key does not
exist within the _dict attribute. I decided on zero rather than on
None due to the fact that calling describe_table on the newly created
table returns zero for both of these attributes.