Refactor boto test organization
The boto tests have been split into two main directories,
unit and integration.  The unit test is intended for test cases
that do not talk to any remote services, while the integration
tests are for any testcases that talk to remote services and
require valid AWS credentials.

I've also updated the tests.py runner to be a lightweight wrapper
to nosetests.  This makes it easier for people to contribute
new tests and ensure they will be run as well as choosing which
tests you want to run.

While splitting the tests and running the tests, there are small
number of minor bug fixes I made (with appropriate tests included).