add resubmit method to BatchList
Resubmit automatically create new batches to get the next set of
results from DynamoDB.

The new request is infered from the ``UnprocessedKeys`` field of
the answer.

If this field was empty, ``None`` is returned instead of the result
set. I choosed ``None`` over letting ``DynamoDBValidationError``
normal behavior propagate because the request is supposed clean at
this point and it allows the use of a clean loop