Add long polling support to SQS
This change adds the wait_time_seconds param to the sqs module.
When receiving a message on an SQS queue, you can now
specify how long to block before returning without any
messages.  This also required a bump of the API version
to 2012-11-05.

I've added some integration tests that demonstrate this feature.