Fix multipart uploads too big for default part size
The default part size of 4 MB is too small for any files over 40,000 MB, and there's no checking to ensure that a part size passed in is big enough.

I added a check to make sure the default part size is big enough. If it's not, the minimum part size needed is found and then it figures what what the smallest part size allowed by Amazon's API is and uses that instead.