Support Expiration date in bucket lifecycle object.
Changed the expiration parameter to be similar to the transition
parameter. It is now an object. The methods still take an
integer number of days but any code that looks directly at
the value of rule.expiration and expects it to be an integer
will now fail :-( Not sure how to make the transition smoother.

Added a unit test to test multiple rules together.