Convert an SQS queue policy to a correct format before modifying it
queue.get_attributes('Policy') in subscribe_sqs_queue method returns
a dictionary object which only has 'Policy' as a key. This is not an
expected format in subscribe_sqs_queue and will cause 400 "Bad request"
if specified queue has any policy. So converting the dictionary object
to a correct format before modifying it will avoid it.