Correct layer1 scan documentation for limit param
"Limit - The maximum number of items to evaluate (not necessarily the
number of matching items). If Amazon DynamoDB processes the number of
items up to the limit while processing the results, it stops and returns
the matching values up to that point, and a LastEvaluatedKey to apply in
a subsequent operation to continue retrieving items. Also, if the
scanned data set size exceeds 1MB before Amazon DynamoDB reaches this
limit, it stops the scan and returns the matching values up to the
limit, and a LastEvaluatedKey to apply in a subsequent operation to
continue the scan."

http://docs.amazonwebservices.com/amazondynamodb/latest/developerguide/API_Scan.html