Add support for website bucket redirect rules
This is a backwards compatible change that updates the
``configure_website`` method on the ``Bucket`` class with
extra parameters.

I've extracted the website configuration into it's own class,
similar to the CORS and Lifecycle configurations, but only
used it internally for the implementation of configure_website.
We could also add a method on Bucket that takes a WebsiteConfiguration
object directly, but I've not added that.

This is a direct mapping to the API as specified in:
http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketPUTwebsite.html

I've added unittests and integration tests for these changes,
and the existing integration test for website configuration still
passes.