Remove fp seek check if fp cannot seek
Some file-like objects do not support seeking. If the size and md5 of the
underlying data is already known in advance, it is possible and desirable to
read the stream only when we are actually sending the bytes to the storage
service. The current fp.seek calls are simply a sanity check that a user didn't
pass in an already-advanced fp, and can safely be avoided if the fp does not
support seeking.