Adjust the output of base64.encodestring in p3k
Apparently a string of the form "%s:%s" % (x, y), when y is a byte
literal, will be formatted as "this-is-x:b'this-is-y'", which
obviously AWS isn't expecting.  So this patch is further insanity
to ensure that the output of base64.encodestring is a utf-8 encoded
string, not a byte literal.