Initial integration tests for idempotent subscribe
For https://github.com/boto/boto/pull/1414

There is occasionally a failure on the first assert in
'test_idempotent_subscribe' where Amazon hasn't yet propogated the
policy change when we do the get_attributes after the first subscribe,
so it still has 0 statements in it.  I'm not sure what the best way to
deal with this is.