Adjust part size as needed in create_archive_from_file
Fixes #1478, original patch suggested by @dminkovsky.
If we have a large enough file such that the default part size
is not sufficient to stay within the maximum number of parts
allowed by glacier, we automatically adjust the part size
to the minimum required.  We were previously calculating this
value but not actually doing anything with it.