Return the x-amz-restore header with GET KEY and fix provider prefix.
Currently x-amz-restore is only supported by HEAD KEY but GET KEY also
returns the header when it has been successfully restored and the
restored key is awaiting expiry. Fixed the provider 'x-amz-' prefix
part so that it's consistent with other standard headers. Previously,
it was hardcoded.

This patch is hard to automate a test for as transitioning to GLACIER
and restoring back from GLACIER both take time. This change has been
manually tested.