Align volume delete on terminate with AWS defaults
Updated the BlockDeviceType to have a default to terminate volumes on
instance termination, the same as the AWS default as documented in these
pages:
  http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/block-device-mapping-concepts.html
  http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RegisterImage.html

Also updated register_image to allow for users to explicitly set their
delete on terminate preference when creating an image from a snapshot.

Updated tests to work with the new default, and created new tests to
assert that the default behavior is in place.