Remove auto tag from create_snapshot
Snapshotting does not imply tagging rights. This regression causes UnauthorizedOperation's to be thrown when tagging fails, which imply a failure of the snapshot when a failure of the snapshot did not occur. This tagging feature should be implemented as a new method that does both, or new parameter that is opt-in, not a backward-incompatible change of the existing behavior of the method.