Make sure integer values in rds parameter are not based on variables
Some RDS parameter group values may be based on variables. For example,
by default the innodb_buffer_pool_size parameter has a default value of
'{DBInstanceClassMemory*3/4}' when the group is created. Trying to even
fetch the value of this parameter results in a traceback:

  File "../site-packages/boto/rds/parametergroup.py", line 184, in get_value
    self._set_integer_value(self._value)
  File "../site-packages/boto/rds/parametergroup.py", line 146, in _set_integer_value
    value = int(value)
ValueError: invalid literal for int() with base 10: '{DBInstanceClassMemory/12582880}'

This patch corrects this by checking if the value is a basestring, and
only converting it to an integer if it does not start with a '{' as well
as checks for the same '{' when trying to save the integer value.