Add namespace to AccessControlPolicy xml representation
Scality implements S3 which is for instance used by GreenQloud to
provide an S3 interface to their Storage Qloud. Scality's
implementation requires the AccessControlPolicy to include a
namespace attribute to work properly. This xmls namespace is
retrieved from the parsed AccessControlPolicy.

This is also what the example in the Amazon S3 docs does:
http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketPUTacl.html

This commit adds the namespace attribute to the xml representation
of the Policy object so that it can be used to set acl grants thus
providing support for Scality storage clouds, like the GreenQloud
one, in boto.

Thanks to Tryggvi Lárusson, CTO of GreenQloud, for identifying this.