remove obsolete md5 import routine
The old md5 module was deprecated since Python 2.5, and we even didn't
support that version already. The current implementation in
boto/utils.py will not even work under Python 2.4 because of the name
"md5" imported from md5 module was replaced by the md5 module in the
following routine, which will cause errors.