BUG: Prevents os.remove from masking an error with open in get_contents_to_filename
When exceptions were raised because open failed, the exception handler
would call os.remove(filename) which would raise an exception saying
that the file does not exist. This would mask the exception that open
raised.

By checking for the existence of the file on disk prior to removing it,
these types of exceptions should no longer be masked.