Parse unicode query string properly in Python 2
This fixes incorrect signing in sigv4 when using
non-ascii characters in query string in Python 2
for S3 requests. The query string parsed for
sigv4 calculation is done incorrectly in Python 2
when given a Unicode query with URL-encoded
non-ascii characters. This is because
urllib.parse.parse_qs in Python 2 doesn't do
decoding of URL-encoded %HH bytes.