Fix line-ending bug in auth variables.
When adding environment variables to a bash script with Windows line endings, environment variables can, in a UNIX environment, add those line endings into the value of the key. Given standard behavior of bash variables (and shell variables in general), I propose this change to better reduce confusion. This is not as uncommon as it sounds- the Amazon default key CSV files have CRLF endings, resulting in sadness when a student of mine edits the file in place to look like a shell script.