Location MetaClass and validation
Having a metaclass for Location (i.e. MetaLocation) allows to store the
locations in a dict, rather than using the static class variables. This
way, it is possible to check 'if ... in Location' for example.

It is used to validate the user input against the contents of the region
dict; in case the user input is bogus, or in case the region dict is not
up to date.