Fix key.set_contents_from_file() if fp.name is int
set_contents_from_file uses fp.name as a path if it exists.

However, fp.name may be a file descriptor (an integer), which is not suitable
as a path, and leads to an error when guess_type is called.

See https://docs.python.org/3/library/io.html#io.FileIO