Fix SQS binary message decoding
Encoding and decoding of messages stopped working in
version 2.32.0 with the introduction of Python 3 support.
Encoding of binary SQS messages was fixed in 2.36.0 but
decoding of those messages was still not transparent to
the user as it had been prior to 2.32.0. This fixes that
by attempting to utf-8 decode the base64 decoded data. If
this does not succeed, the base64 decoded data is returned.