Read response body just before throwing exception in _mexe
Previously if there was a retry_handler configured for _mexe and it
retried the maximum number of times, when the BotoServerError was thrown
the response body would not be populated because it was only being read
if the AWS error was in the 500 range. This made it hard to debug what
the exact error code being returned was.

This change moves the reading of the response body to just before the
BotoServerError is thrown so that the body is read in all cases. We also
added a test case that exposed the original problem.

Miles O'Connell (@milesoc) paired with me on this change.