get_all_reservations: check kwarg with `is not None`
this bit us when we expected `get_all_instances` to return no instances if we passed it `instance_ids=[]`. we were surprised when it returned us all of our instances instead. :P

i know boto is currently in maintenance mode in favor of boto3, so up to you if you want to merge.

cc @justinwang189