Fix travis CI builds for PY3
Tracking issue here: https://github.com/travis-ci/travis-ci/issues/5246.
What's happening is that the global boto config on the travis CI
systems is telling boto to load a plugin that is not compatible with
python3.

In the meantime, our workaround is to set the BOTO_CONFIG to some bogus
value so we don't load the global config.