Add better error messages when we can't retrieve creds
When boto retrieves credentials from the instance metadata
service, there appear to be times when we're just returning an empty
string back to the user.  When this happens, we need to raise
a much more clear error message about what's going on.

There's a rare chance that we're actually seeing a 200 OK response
with text data that isn't the credentials in JSON format.  If that's
the case we'll need another change that counts that condition as a
failed attempt.

Fixes #3378.