Set self.etag to lowercase for md5 comparison
I was working on using boto against an S3-compatible service, but the service returns etags as uppercase, so it fails to match md5 comparison. I wrapped self.etag in `str()` in the event it is `None`, so `lower()` doesn't fail -- that seems like the the right thing to do. I could also do `if self.etag and self.etag.lower()`. I'm not sure if you'd have a preference. :)