Check whether query_auth is true before adding x-amz-security-token for S3 URL generation.
When the S3 component of Boto was used on an EC2 instance that was
granted permissions via an IAM role, it attempted to append
`x-amz-security-token` key and value to the URL generated via
the `generate_url` method on the available connection, even though
`query_auth=False` was explicitly being set.

Includes a simple test case, and the test passes on all interpreters
specified in the current tox configuration.

Fixes #2043
Refs. #2044