Removes documented support for Python 3 with S3
The S3 service does not entirely work with Python 3 (as pointed out in issue #3561). This issue issue has been fixed by PR #2718 long before it was even reported, yet that PR has not been merged yet. Until that PR is merged, the boto documentation should not list Python 3 support for the S3 service, as it does not work in Python 3.