Add ugly patch to allow values from a range given by AWS
As background, the 'allowed_values' field is normally a list, i.e.
['a','b','c'] or [1,2,3]. However updates to AWS changed this to also
permit a range of values (particularly for floats/ints). This means that
you may receive a 'allowed_values' field of: ['0-1000']. The logic that
checks whether 'my_value' in ['0-1000'] will always fail, because you
would have to literally pass '0-1000' instead of your desired value.

While the true solution here is to actually put in logic to locate this
case and check whether your value falls within the range given, instead
this patch just says that if there is only one option (likely to only
occur in this type of case) we'll just allow you to set the value, at
which point the API itself will error if you didn't follow the
guidelines of the API.