KmsConnection#make_request - correctly set custom Host header
This fixes a bug when using a `KmsConnection` with a custom host
value - for example, [FIPS endpoints](https://aws.amazon.com/compliance/fips/).

Previously, kms would set the host header to the region endpoint,
regardless of the actual host that would be used. That resulted in
an `InvalidSignatureException` when making any call.